requirements.txt
aniso8601==7.0.0
apispec==6.6.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docopt==0.6.2
dotenv==0.0.5
email_validator==2.1.1
executing==2.0.1
fastjsonschema==2.20.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-GraphQL==2.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-smorest==0.44.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
graphene==2.1.9
graphene-sqlalchemy
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==3.0.3
idna==3.7
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.2
promise==2.3
prompt_toolkit==3.0.47
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==306
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
Rx==1.6.3
singledispatch==3.7.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.1
urllib3==2.2.2
wcwidth==0.2.13
webargs==8.4.0
webencodings==0.5.1
Werkzeug==3.0.3
wtf==0.1
WTForms==3.1.2
yarg==0.1.9
